search: split mod.rs into index, query, activation, overlay modules
The single mod.rs mixed index construction, the query pipeline,
activation state tracking, and draft overlay logic across ~4100
lines. Each concern now lives in its own file under search/.

mod.rs retains the public entry points and shared infrastructure.
